+++ frontend/src/common/components/ListItem.tsx
@@ -39,7 +39,7 @@ export const ListItem: React.FC<ListItemProps> = props => {
                     }
                     } />
                 </span>
-                <button type="button" title='Execute' is="paper-icon-button-light" className="paper-icon-button-light icon-button-conditionalfocuscolor" onClick={() => props.onExecuteClick()}><i className="md-icon sp-icon">play_arrow</i></button>
+                <button type="button" disabled={props.playList.updateType === 'Live' ? true : undefined} title={props.playList.updateType === 'Live' ? "Cannot execute 'Live' update types" : 'Execute'} is="paper-icon-button-light" className="paper-icon-button-light icon-button-conditionalfocuscolor" onClick={() => props.onExecuteClick()}><i className="md-icon sp-icon">play_arrow</i></button>
                 <button type="button" title='Details' is="paper-icon-button-light" className="paper-icon-button-light icon-button-conditionalfocuscolor" onClick={() => props.onViewClick()}><i className="md-icon sp-icon">info</i></button>
                 <button type="button" title='Edit' is="paper-icon-button-light" className="paper-icon-button-light icon-button-conditionalfocuscolor" onClick={() => props.onEditClick()}><i className="md-icon sp-icon">edit</i></button>
                 {props.playList.smartType == "Playlist" && (